backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
requests
groq
gtts
python-multipart
pypdf
python-docx
